genostream-analytics/requirements.txt
fastapi
uvicorn[standard]
clickhouse-driver[numpy]
pandas
python-dotenv
pydantic-settings